Initial ASC commit
The code compiles, and is representative of what I want my hybrid guideline-ASC gametype to play like. Lots of unneeded code (and one block of code that crashes on asc, it's only for optimization of a bot's first move though) is cut out, but some still remains (cutouts, which super aren't needed in immobile all-spins).